+++ src/main/java/chess/domain/game/ChessException.java
@@ -1,4 +1,4 @@
-package chess.exception;
+package chess.domain.game;
 
 public class ChessException extends RuntimeException {
     public ChessException(String message) {
